Allow non-int pixel locations

+++ factor/directions.py
@@ -559,12 +559,12 @@ def thiessen(directions_list, field_ra_deg, field_dec_deg, bounds_scale=0.5,
                 d.is_patch = True # set patch flag to ensure facet is not included in mosaic
                 sx, sy = radec2xy([d.ra], [d.dec], refRA=midRA, refDec=midDec)
                 patch_width = d.cal_size_deg * 1.2 / 0.066667 # size of patch in pixels
-                x0 = int(sx[0] - patch_width / 2.0)
-                y0 = int(sy[0] - patch_width / 2.0)
+                x0 = sx[0] - patch_width / 2.0
+                y0 = sy[0] - patch_width / 2.0
                 poly = [np.array([x0, y0]),
-                        np.array([x0, int(y0+patch_width)]),
-                        np.array([int(x0+patch_width), int(y0+patch_width)]),
-                        np.array([int(x0+patch_width), y0])]
+                        np.array([x0, y0+patch_width]),
+                        np.array([x0+patch_width, y0+patch_width]),
+                        np.array([x0+patch_width, y0])]
 
         poly = np.vstack([poly, poly[0]])
         ra, dec = xy2radec(poly[:, 0], poly[:, 1], midRA, midDec)
